requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.7
ifcopenshell==0.7.10
numpy==1.24.3
pandas==2.0.3
jinja2==3.1.2
aiofiles==23.2.1
pydantic==2.5.0
toolregistry
# Phoenix tracing dependencies
arize-phoenix>=4.0.0
openinference-instrumentation-openai>=0.1.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
instructor>=1.0.0 